CreateProduct page with optional fields
Tested and working on my machine. I plan on updating this later with dropdown list options for the category and location fields, but for now this is working. *FYI THIS IS A BRANCH OF THE CREATEACCOUNT BRANCH: It'll be quicker and simpler to just merge this branch once it is reviewed*