Working NavBar
The bar shows links to pages dependent on the user session token. The only page links visible without login are /Login and /CreateAccount, which themselves are not visible if the user IS logged in.
Feel free to change the link names if you want.
I've also been playing around with the CSS styling of the links because I don't like how they're borderless and flow into each other. This version is using the default styling though.