Working CSV Download Feature
You can download the Product list as a comma-separated CSV file with a header.
This feature works with the query based filtering. The downloaded CSV will correspond to whatever the displayed result of your query was, unless your result was an empty list.
When you download without a query, or when an empty list is displayed, the download will give you the full product list as a csv.