Account settings feature working
I've bugtested this a lot and am certain it is good, mostly because it had a lot of errors that took a long time to fix, so I've covered them all.
But regardless PLEASE test this feature before reviewing.

This feature comes with the Account Settings page and an AccountService class. I planned on implementing this service class on login and createAccount, but considering how long this feature took to get working I won't be doing that this week.